Add guardrails ai (#9)
* add git ignore

* feat(docker): add Dockerfile for containerized deployment

Introduced a Dockerfile to streamline the deployment process of the FastAPI application. The Dockerfile sets up a Python 3.11 slim environment, installs necessary dependencies, and configures the application to run with Guardrails AI for drug mention detection and web sanitization. Updated README with instructions for building the Docker image and added new endpoints for drug mention detection and web sanitization.